+++ cmd/quadrepl/src/main.cc
@@ -53,7 +53,7 @@ static const ReferenceEntry g_reference[] = {
 #undef KEYWORD
 };
 
-static const char* const g_keywords[] = {"fn", "pub", "inline", "if", "else", "for", "loop", "break", "continue",
+static const char* const g_keywords[] = {"fn", "pub", "inline", "if", "else", "for", "loop", "while", "break", "continue",
 		"return", "use", "struct", "packed", "enum", "const", "var", "defer", "switch", "case", "test", "type", "as",
 		"null", "true", "false", "Ok", "Err"};