+++ lib/qc/src/ast_printer.cc
@@ -90,6 +90,8 @@ namespace Qd {
 			return "ArrayIndex";
 		case IAstNode::Type::LOOP_STATEMENT:
 			return "LoopStatement";
+		case IAstNode::Type::WHILE_STATEMENT:
+			return "WhileStatement";
 		case IAstNode::Type::IMPORT_STATEMENT:
 			return "ImportStatement";
 		case IAstNode::Type::GLOBAL_VAR_DECLARATION: